Refactor ChatViewModel into managers
