+++ bitchat/ViewModels/ChatViewModel.swift
@@ -145,6 +145,8 @@ class ChatViewModel: ObservableObject, BitchatDelegate {
     private var nostrRelayManager: NostrRelayManager?
     private var messageRouter: MessageRouter?
     private var peerManager: PeerManager?
+    private lazy var commandProcessor = CommandProcessor(viewModel: self)
+    private lazy var privateChatManager = PrivateChatManager(viewModel: self)
     private let userDefaults = UserDefaults.standard
     private let nicknameKey = "bitchat.nickname"
     
@@ -573,94 +575,18 @@ class ChatViewModel: ObservableObject, BitchatDelegate {
                 self?.markPrivateMessagesAsRead(from: peerID)
             }
         }
-    }
-    
     private func isPeerBlocked(_ peerID: String) -> Bool {
-        // Check if we have the public key fingerprint for this peer
-        if let fingerprint = peerIDToPublicKeyFingerprint[peerID] {
-            return SecureIdentityStateManager.shared.isBlocked(fingerprint: fingerprint)
-        }
-        
-        // Try to get fingerprint from mesh service
-        if let fingerprint = meshService.getPeerFingerprint(peerID) {
-            return SecureIdentityStateManager.shared.isBlocked(fingerprint: fingerprint)
-        }
-        
-        return false
+        privateChatManager.isPeerBlocked(peerID)
     }
-    
-    // Helper method to find current peer ID for a fingerprint
+
     private func getCurrentPeerIDForFingerprint(_ fingerprint: String) -> String? {
-        // Search through all connected peers to find the one with matching fingerprint
-        for peerID in connectedPeers {
-            if let mappedFingerprint = peerIDToPublicKeyFingerprint[peerID],
-               mappedFingerprint == fingerprint {
-                return peerID
-            }
-        }
-        return nil
+        // Wrapper for PrivateChatManager
+        return nil // unused after refactor
     }
     
     // Helper method to update selectedPrivateChatPeer if fingerprint matches
     private func updatePrivateChatPeerIfNeeded() {
-        guard let chatFingerprint = selectedPrivateChatFingerprint else { return }
-        
-        // Find current peer ID for the fingerprint
-        if let currentPeerID = getCurrentPeerIDForFingerprint(chatFingerprint) {
-            // Update the selected peer if it's different
-            if let oldPeerID = selectedPrivateChatPeer, oldPeerID != currentPeerID {
-                SecureLogger.log("📱 Updating private chat peer from \(oldPeerID) to \(currentPeerID)", 
-                                category: SecureLogger.session, level: .debug)
-                
-                // Migrate messages from old peer ID to new peer ID
-                if let oldMessages = privateChats[oldPeerID] {
-                    if privateChats[currentPeerID] == nil {
-                        privateChats[currentPeerID] = []
-                    }
-                    privateChats[currentPeerID]?.append(contentsOf: oldMessages)
-                    // Sort by timestamp
-                    privateChats[currentPeerID]?.sort { $0.timestamp < $1.timestamp }
-                    // Remove duplicates
-                    var seen = Set<String>()
-                    privateChats[currentPeerID] = privateChats[currentPeerID]?.filter { msg in
-                        if seen.contains(msg.id) {
-                            return false
-                        }
-                        seen.insert(msg.id)
-                        return true
-                    }
-                    trimPrivateChatMessagesIfNeeded(for: currentPeerID)
-                    privateChats.removeValue(forKey: oldPeerID)
-                }
-                
-                // Migrate unread status
-                if unreadPrivateMessages.contains(oldPeerID) {
-                    unreadPrivateMessages.remove(oldPeerID)
-                    unreadPrivateMessages.insert(currentPeerID)
-                }
-                
-                selectedPrivateChatPeer = currentPeerID
-                
-                // Schedule UI update for encryption status change
-                DispatchQueue.main.async { [weak self] in
-                    self?.scheduleUIUpdate()
-                }
-                
-                // Also refresh the peer list to update encryption status
-                Task { @MainActor in
-                    peerManager?.updatePeers()
-                }
-            } else if selectedPrivateChatPeer == nil {
-                // Just set the peer ID if we don't have one
-                selectedPrivateChatPeer = currentPeerID
-                DispatchQueue.main.async { [weak self] in
-                    self?.scheduleUIUpdate()
-                }
-            }
-            
-            // Clear unread messages for the current peer ID
-            unreadPrivateMessages.remove(currentPeerID)
-        }
+        privateChatManager.updatePrivateChatPeerIfNeeded()
     }
     
     // MARK: - Message Sending
@@ -676,7 +602,7 @@ class ChatViewModel: ObservableObject, BitchatDelegate {
         // Check for commands
         if content.hasPrefix("/") {
             Task { @MainActor in
-                handleCommand(content)
+                commandProcessor.handle(content)
             }
             return
         }
@@ -830,149 +756,13 @@ class ChatViewModel: ObservableObject, BitchatDelegate {
     /// - Parameter peerID: The peer's ID to start chatting with
     /// - Note: Switches the UI to private chat mode and loads message history
     @MainActor
+    @MainActor
     func startPrivateChat(with peerID: String) {
-        // Safety check: Don't allow starting chat with ourselves
-        if peerID == meshService.myPeerID {
-            SecureLogger.log("⚠️ Attempted to start private chat with self, ignoring", 
-                            category: SecureLogger.session, level: .warning)
-            return
-        }
-        
-        // Try to get nickname from mesh first, then from peer list
-        let peerNickname = meshService.getPeerNicknames()[peerID] ?? 
-                          peerIndex[peerID]?.displayName ?? 
-                          "unknown"
-        
-        // Check if the peer is blocked
-        if isPeerBlocked(peerID) {
-            let systemMessage = BitchatMessage(
-                sender: "system",
-                content: "cannot start chat with \(peerNickname): user is blocked.",
-                timestamp: Date(),
-                isRelay: false
-            )
-            messages.append(systemMessage)
-            return
-        }
-        
-        // Check if this is a moon peer (we favorite them but they don't favorite us) AND they're offline
-        // Only require mutual favorites for offline Nostr messaging
-        if let peer = peerIndex[peerID],
-           peer.isFavorite && !peer.theyFavoritedUs && !peer.isConnected && !peer.isRelayConnected {
-            let systemMessage = BitchatMessage(
-                sender: "system",
-                content: "cannot start chat with \(peerNickname): mutual favorite required for offline messaging.",
-                timestamp: Date(),
-                isRelay: false
-            )
-            messages.append(systemMessage)
-            return
-        }
-        
-        // Trigger handshake if we don't have a session yet
-        let sessionState = meshService.getNoiseSessionState(for: peerID)
-        switch sessionState {
-        case .none, .failed:
-            // Initiate handshake when opening PM
-            meshService.triggerHandshake(with: peerID)
-        default:
-            break
-        }
-        
-        selectedPrivateChatPeer = peerID
-        // Also track by fingerprint for persistence across reconnections
-        selectedPrivateChatFingerprint = peerIDToPublicKeyFingerprint[peerID]
-        unreadPrivateMessages.remove(peerID)
-        
-        // Check if we need to migrate messages from an old peer ID
-        // This happens when peer IDs change between sessions
-        if privateChats[peerID] == nil || privateChats[peerID]?.isEmpty == true {
-            
-            // Get the fingerprint for this peer
-            let currentFingerprint = getFingerprint(for: peerID)
-            
-            // Look for messages under other peer IDs with the same fingerprint
-            var migratedMessages: [BitchatMessage] = []
-            var oldPeerIDsToRemove: [String] = []
-            
-            for (oldPeerID, messages) in privateChats {
-                if oldPeerID != peerID {
-                    // Check if this old peer ID has the same fingerprint (same actual peer)
-                    let oldFingerprint = peerIDToPublicKeyFingerprint[oldPeerID]
-                    
-                    if let currentFp = currentFingerprint,
-                       let oldFp = oldFingerprint,
-                       currentFp == oldFp {
-                        // Same peer with different ID - migrate all messages
-                        migratedMessages.append(contentsOf: messages)
-                        oldPeerIDsToRemove.append(oldPeerID)
-                        
-                        SecureLogger.log("📦 Migrating \(messages.count) messages from old peer ID \(oldPeerID) to \(peerID) based on fingerprint match", 
-                                        category: SecureLogger.session, level: .info)
-                    } else if currentFingerprint == nil || oldFingerprint == nil {
-                        // Fallback: use nickname matching only if we don't have fingerprints
-                        // This is less reliable but handles legacy data
-                        let messagesWithPeer = messages.filter { msg in
-                            // Message is FROM the peer to us
-                            (msg.sender == peerNickname && msg.sender != nickname) ||
-                            // OR message is FROM us TO the peer
-                            (msg.sender == nickname && (msg.recipientNickname == peerNickname || 
-                             // Also check if this was a private message in a chat that only has us and one other person
-                             (msg.isPrivate && messages.allSatisfy { m in 
-                                 m.sender == nickname || m.sender == peerNickname 
-                             })))
-                        }
-                        
-                        if !messagesWithPeer.isEmpty {
-                            // Check if ALL messages in this chat are between us and this peer
-                            let allMessagesAreWithPeer = messages.allSatisfy { msg in
-                                (msg.sender == peerNickname || msg.sender == nickname) &&
-                                (msg.recipientNickname == nil || msg.recipientNickname == peerNickname || msg.recipientNickname == nickname)
-                            }
-                            
-                            if allMessagesAreWithPeer {
-                                // This entire chat history likely belongs to this peer, migrate it all
-                                migratedMessages.append(contentsOf: messages)
-                                oldPeerIDsToRemove.append(oldPeerID)
-                                
-                                SecureLogger.log("📦 Migrating \(messages.count) messages from old peer ID \(oldPeerID) to \(peerID) based on nickname match (no fingerprints available)", 
-                                                category: SecureLogger.session, level: .warning)
-                            }
-                        }
-                    }
-                }
-            }
-            
-            // Remove old peer ID entries that were fully migrated
-            for oldPeerID in oldPeerIDsToRemove {
-                privateChats.removeValue(forKey: oldPeerID)
-                unreadPrivateMessages.remove(oldPeerID)
-            }
-            
-            // Initialize chat history with migrated messages if any
-            if !migratedMessages.isEmpty {
-                privateChats[peerID] = migratedMessages.sorted { $0.timestamp < $1.timestamp }
-                trimPrivateChatMessagesIfNeeded(for: peerID)
-            } else {
-                privateChats[peerID] = []
-            }
-        }
-        
-        _ = privateChats[peerID] ?? []
-        
-        // Send read receipts for unread messages from this peer
-        // Add a small delay to ensure UI has updated
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) { [weak self] in
-            self?.markPrivateMessagesAsRead(from: peerID)
-        }
-        
-        // Also try immediately in case messages are already there
-        markPrivateMessagesAsRead(from: peerID)
+        privateChatManager.startPrivateChat(with: peerID)
     }
-    
+
     func endPrivateChat() {
-        selectedPrivateChatPeer = nil
-        selectedPrivateChatFingerprint = nil
+        privateChatManager.endPrivateChat()
     }
     
     // MARK: - Message Retry Handling
@@ -1328,158 +1118,14 @@ class ChatViewModel: ObservableObject, BitchatDelegate {
     }
     
     @MainActor
+    @MainActor
     func markPrivateMessagesAsRead(from peerID: String) {
-        // Get the nickname for this peer
-        let peerNickname = meshService.getPeerNicknames()[peerID] ?? ""
-        
-        // First ensure we have the latest messages (in case of migration)
-        if let messages = privateChats[peerID], !messages.isEmpty {
-        } else {
-            
-            // Look through ALL private chats to find messages from this nickname
-            for (_, chatMessages) in privateChats {
-                let relevantMessages = chatMessages.filter { msg in
-                    msg.sender == peerNickname && msg.sender != nickname
-                }
-                if !relevantMessages.isEmpty {
-                }
-            }
-        }
-        
-        guard let messages = privateChats[peerID], !messages.isEmpty else { 
-            return 
-        }
-        
-        
-        // Find messages from the peer that haven't been read yet
-        var readReceiptsSent = 0
-        for (_, message) in messages.enumerated() {
-            // Only send read receipts for messages from the other peer (not our own)
-            // Check multiple conditions to ensure we catch all messages from the peer
-            let isOurMessage = message.sender == nickname
-            let isFromPeerByNickname = !peerNickname.isEmpty && message.sender == peerNickname
-            
-            // Check if message is from this peer by comparing IDs
-            // Handle both Noise and Nostr keys
-            var isFromPeerByID = false
-            if let msgSenderID = message.senderPeerID {
-                // Direct match
-                isFromPeerByID = msgSenderID == peerID
-                
-                // If no direct match, check if we're comparing Nostr keys
-                if !isFromPeerByID {
-                    // Check if the peerID has a favorite entry with a Nostr key that matches
-                    if let noiseKey = Data(hexString: peerID),
-                       let favoriteStatus = FavoritesPersistenceService.shared.getFavoriteStatus(for: noiseKey),
-                       let nostrKey = favoriteStatus.peerNostrPublicKey {
-                        isFromPeerByID = msgSenderID == nostrKey
-                    }
-                    
-                    // getFavoriteStatusByNostrKey not implemented
-                    // Or check if the message sender ID maps to this peer's Noise key
-                    // if !isFromPeerByID,
-                    //    let senderFavorite = FavoritesPersistenceService.shared.getFavoriteStatusByNostrKey(msgSenderID),
-                    //    senderFavorite.peerNoisePublicKey.hexEncodedString() == peerID {
-                    //     isFromPeerByID = true
-                    // }
-                }
-            }
-            
-            let isPrivateToUs = message.isPrivate && message.recipientNickname == nickname
-            
-            // This is a message FROM the peer if it's not from us AND (matches nickname OR peer ID OR is private to us)
-            let isFromPeer = !isOurMessage && (isFromPeerByNickname || isFromPeerByID || isPrivateToUs)
-            
-            if message.id == message.id { // Always true, for debugging
-            }
-            
-            if isFromPeer {
-                if let status = message.deliveryStatus {
-                    switch status {
-                    case .sent, .delivered:
-                        // Create and send read receipt for sent or delivered messages
-                        // Check if we've already sent a receipt for this message
-                        if !sentReadReceipts.contains(message.id) {
-                            let receipt = ReadReceipt(
-                                originalMessageID: message.id,
-                                readerID: meshService.myPeerID,
-                                readerNickname: nickname
-                            )
-                            
-                            // Send read receipt to the message sender
-                            // Use the message's senderPeerID if available (for Nostr messages)
-                            // Otherwise use the current peerID
-                            let recipientID = message.senderPeerID ?? peerID
-                            
-                            // Check if message was delivered via Nostr
-                            var originalTransport: String? = nil
-                            if let status = message.deliveryStatus {
-                                switch status {
-                                case .delivered(let transport, _):
-                                    originalTransport = transport
-                                default:
-                                    break
-                                }
-                            }
-                            
-                            sendReadReceipt(receipt, to: recipientID, originalTransport: originalTransport)
-                            sentReadReceipts.insert(message.id)
-                            readReceiptsSent += 1
-                        } else {
-                        }
-                    case .read:
-                        // Already read, no need to send another receipt
-                        break
-                    default:
-                        // Message not yet delivered, can't mark as read
-                        break
-                    }
-                } else {
-                    // No delivery status - this might be an older message
-                    // Send read receipt anyway for backwards compatibility
-                    if !sentReadReceipts.contains(message.id) {
-                        let receipt = ReadReceipt(
-                            originalMessageID: message.id,
-                            readerID: meshService.myPeerID,
-                            readerNickname: nickname
-                        )
-                        // Use the message's senderPeerID if available (for Nostr messages)
-                        let recipientID = message.senderPeerID ?? peerID
-                        
-                        // For backwards compatibility, check if this might be a Nostr message
-                        // by checking if the sender has a Nostr key
-                        let receiptToSend = receipt
-                        let recipientToSend = recipientID
-                        Task { @MainActor in
-                            var originalTransport: String? = nil
-                            if let noiseKey = Data(hexString: recipientToSend),
-                               let favoriteStatus = FavoritesPersistenceService.shared.getFavoriteStatus(for: noiseKey),
-                               favoriteStatus.peerNostrPublicKey != nil,
-                               self.meshService.getPeerNicknames()[recipientToSend] == nil {
-                                // Peer has Nostr key and is not on mesh, likely received via Nostr
-                                originalTransport = "nostr"
-                            }
-                            
-                            self.sendReadReceipt(receiptToSend, to: recipientToSend, originalTransport: originalTransport)
-                        }
-                        sentReadReceipts.insert(message.id)
-                        readReceiptsSent += 1
-                    } else {
-                    }
-                }
-            } else {
-            }
-        }
-        
+        privateChatManager.markPrivateMessagesAsRead(from: peerID)
     }
     
     func getPrivateChatMessages(for peerID: String) -> [BitchatMessage] {
-        let messages = privateChats[peerID] ?? []
-        if !messages.isEmpty {
-        }
-        return messages
+        privateChatManager.getPrivateChatMessages(for: peerID)
     }
-    
     func getPeerIDForNickname(_ nickname: String) -> String? {
         let nicknames = meshService.getPeerNicknames()
         return nicknames.first(where: { $0.value == nickname })?.key
@@ -2095,19 +1741,9 @@ class ChatViewModel: ObservableObject, BitchatDelegate {
             messages.removeFirst(removeCount)
         }
     }
-    
     private func trimPrivateChatMessagesIfNeeded(for peerID: String) {
-        if let count = privateChats[peerID]?.count, count > maxMessages {
-            let removeCount = count - maxMessages
-            privateChats[peerID]?.removeFirst(removeCount)
-        }
+        privateChatManager.trimPrivateChatMessagesIfNeeded(for: peerID)
     }
-    
-    private func addMessageToBatch(_ message: BitchatMessage) {
-        pendingMessages.append(message)
-        scheduleBatchFlush()
-    }
-    
     private func addPrivateMessageToBatch(_ message: BitchatMessage, for peerID: String) {
         if pendingPrivateMessages[peerID] == nil {
             pendingPrivateMessages[peerID] = []
@@ -2387,506 +2023,10 @@ class ChatViewModel: ObservableObject, BitchatDelegate {
     // MARK: - BitchatDelegate Methods
     
     // MARK: - Command Handling
-    
-    /// Processes IRC-style commands starting with '/'.
-    /// - Parameter command: The full command string including the leading slash
-    /// - Note: Supports commands like /nick, /msg, /who, /slap, /clear, /help
     @MainActor
     private func handleCommand(_ command: String) {
-        let parts = command.split(separator: " ")
-        guard let cmd = parts.first else { return }
-        
-        switch cmd {
-        case "/m", "/msg":
-            if parts.count > 1 {
-                let targetName = String(parts[1])
-                // Remove @ if present
-                let nickname = targetName.hasPrefix("@") ? String(targetName.dropFirst()) : targetName
-                
-                // Find peer ID for this nickname
-                if let peerID = getPeerIDForNickname(nickname) {
-                    startPrivateChat(with: peerID)
-                    
-                    // If there's a message after the nickname, send it
-                    if parts.count > 2 {
-                        let messageContent = parts[2...].joined(separator: " ")
-                        sendPrivateMessage(messageContent, to: peerID)
-                    } else {
-                        let systemMessage = BitchatMessage(
-                            sender: "system",
-                            content: "started private chat with \(nickname)",
-                            timestamp: Date(),
-                            isRelay: false
-                        )
-                        messages.append(systemMessage)
-                    }
-                } else {
-                    let systemMessage = BitchatMessage(
-                        sender: "system",
-                        content: "user '\(nickname)' not found. they may be offline or using a different nickname.",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(systemMessage)
-                }
-            } else {
-                let systemMessage = BitchatMessage(
-                    sender: "system",
-                    content: "usage: /m @nickname [message] or /m nickname [message]",
-                    timestamp: Date(),
-                    isRelay: false
-                )
-                messages.append(systemMessage)
-            }
-        case "/w":
-            let peerNicknames = meshService.getPeerNicknames()
-            if connectedPeers.isEmpty {
-                let systemMessage = BitchatMessage(
-                    sender: "system",
-                    content: "no one else is online right now.",
-                    timestamp: Date(),
-                    isRelay: false
-                )
-                messages.append(systemMessage)
-            } else {
-                let onlineList = connectedPeers.compactMap { peerID in
-                    peerNicknames[peerID]
-                }.sorted().joined(separator: ", ")
-                
-                let systemMessage = BitchatMessage(
-                    sender: "system",
-                    content: "online users: \(onlineList)",
-                    timestamp: Date(),
-                    isRelay: false
-                )
-                messages.append(systemMessage)
-            }
-        case "/clear":
-            // Clear messages based on current context
-            if let peerID = selectedPrivateChatPeer {
-                // Clear private chat
-                privateChats[peerID]?.removeAll()
-            } else {
-                // Clear main messages
-                messages.removeAll()
-            }
-        case "/hug":
-            if parts.count > 1 {
-                let targetName = String(parts[1])
-                // Remove @ if present
-                let nickname = targetName.hasPrefix("@") ? String(targetName.dropFirst()) : targetName
-                
-                // Check if target exists in connected peers
-                if let targetPeerID = getPeerIDForNickname(nickname) {
-                    // Create hug message
-                    let hugMessage = BitchatMessage(
-                        sender: "system",
-                        content: "🫂 \(self.nickname) hugs \(nickname)",
-                        timestamp: Date(),
-                        isRelay: false,
-                        isPrivate: false,
-                        recipientNickname: nickname,
-                        senderPeerID: meshService.myPeerID
-                    )
-                    
-                    // Send as a regular message but it will be displayed as system message due to content
-                    let hugContent = "* 🫂 \(self.nickname) hugs \(nickname) *"
-                    if selectedPrivateChatPeer != nil {
-                        // In private chat, send as private message
-                        if let peerNickname = meshService.getPeerNicknames()[targetPeerID] {
-                            meshService.sendPrivateMessage("* 🫂 \(self.nickname) hugs you *", to: targetPeerID, recipientNickname: peerNickname)
-                        }
-                    } else {
-                        // In public chat
-                        meshService.sendMessage(hugContent)
-                        messages.append(hugMessage)
-                    }
-                } else {
-                    let errorMessage = BitchatMessage(
-                        sender: "system",
-                        content: "cannot hug \(nickname): user not found.",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(errorMessage)
-                }
-            } else {
-                let usageMessage = BitchatMessage(
-                    sender: "system",
-                    content: "usage: /hug <nickname>",
-                    timestamp: Date(),
-                    isRelay: false
-                )
-                messages.append(usageMessage)
-            }
-            
-        case "/slap":
-            if parts.count > 1 {
-                let targetName = String(parts[1])
-                // Remove @ if present
-                let nickname = targetName.hasPrefix("@") ? String(targetName.dropFirst()) : targetName
-                
-                // Check if target exists in connected peers
-                if let targetPeerID = getPeerIDForNickname(nickname) {
-                    // Create slap message
-                    let slapMessage = BitchatMessage(
-                        sender: "system",
-                        content: "🐟 \(self.nickname) slaps \(nickname) around a bit with a large trout",
-                        timestamp: Date(),
-                        isRelay: false,
-                        isPrivate: false,
-                        recipientNickname: nickname,
-                        senderPeerID: meshService.myPeerID
-                    )
-                    
-                    // Send as a regular message but it will be displayed as system message due to content
-                    let slapContent = "* 🐟 \(self.nickname) slaps \(nickname) around a bit with a large trout *"
-                    if selectedPrivateChatPeer != nil {
-                        // In private chat, send as private message
-                        if let peerNickname = meshService.getPeerNicknames()[targetPeerID] {
-                            meshService.sendPrivateMessage("* 🐟 \(self.nickname) slaps you around a bit with a large trout *", to: targetPeerID, recipientNickname: peerNickname)
-                        }
-                    } else {
-                        // In public chat
-                        meshService.sendMessage(slapContent)
-                        messages.append(slapMessage)
-                    }
-                } else {
-                    let errorMessage = BitchatMessage(
-                        sender: "system",
-                        content: "cannot slap \(nickname): user not found.",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(errorMessage)
-                }
-            } else {
-                let usageMessage = BitchatMessage(
-                    sender: "system",
-                    content: "usage: /slap <nickname>",
-                    timestamp: Date(),
-                    isRelay: false
-                )
-                messages.append(usageMessage)
-            }
-            
-        case "/block":
-            if parts.count > 1 {
-                let targetName = String(parts[1])
-                // Remove @ if present
-                let nickname = targetName.hasPrefix("@") ? String(targetName.dropFirst()) : targetName
-                
-                // Find peer ID for this nickname
-                if let peerID = getPeerIDForNickname(nickname) {
-                    // Get fingerprint for persistent blocking
-                    if let fingerprintStr = meshService.getPeerFingerprint(peerID) {
-                        
-                        if SecureIdentityStateManager.shared.isBlocked(fingerprint: fingerprintStr) {
-                            let systemMessage = BitchatMessage(
-                                sender: "system",
-                                content: "\(nickname) is already blocked.",
-                                timestamp: Date(),
-                                isRelay: false
-                            )
-                            messages.append(systemMessage)
-                        } else {
-                            // Update or create social identity with blocked status
-                            if var identity = SecureIdentityStateManager.shared.getSocialIdentity(for: fingerprintStr) {
-                                identity.isBlocked = true
-                                identity.isFavorite = false  // Remove from favorites if blocked
-                                SecureIdentityStateManager.shared.updateSocialIdentity(identity)
-                            } else {
-                                let blockedIdentity = SocialIdentity(
-                                    fingerprint: fingerprintStr,
-                                    localPetname: nil,
-                                    claimedNickname: nickname,
-                                    trustLevel: .unknown,
-                                    isFavorite: false,
-                                    isBlocked: true,
-                                    notes: nil
-                                )
-                                SecureIdentityStateManager.shared.updateSocialIdentity(blockedIdentity)
-                            }
-                            
-                            // Update local sets for UI
-                            blockedUsers.insert(fingerprintStr)
-                            favoritePeers.remove(fingerprintStr)
-                            
-                            let systemMessage = BitchatMessage(
-                                sender: "system",
-                                content: "blocked \(nickname). you will no longer receive messages from them.",
-                                timestamp: Date(),
-                                isRelay: false
-                            )
-                            messages.append(systemMessage)
-                        }
-                    } else {
-                        let systemMessage = BitchatMessage(
-                            sender: "system",
-                            content: "cannot block \(nickname): unable to verify identity.",
-                            timestamp: Date(),
-                            isRelay: false
-                        )
-                        messages.append(systemMessage)
-                    }
-                } else {
-                    let systemMessage = BitchatMessage(
-                        sender: "system",
-                        content: "cannot block \(nickname): user not found.",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(systemMessage)
-                }
-            } else {
-                // List blocked users
-                if blockedUsers.isEmpty {
-                    let systemMessage = BitchatMessage(
-                        sender: "system",
-                        content: "no blocked peers.",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(systemMessage)
-                } else {
-                    // Find nicknames for blocked users
-                    var blockedNicknames: [String] = []
-                    for (peerID, _) in meshService.getPeerNicknames() {
-                        if let fingerprintStr = meshService.getPeerFingerprint(peerID) {
-                            if blockedUsers.contains(fingerprintStr) {
-                                if let nickname = meshService.getPeerNicknames()[peerID] {
-                                    blockedNicknames.append(nickname)
-                                }
-                            }
-                        }
-                    }
-                    
-                    let blockedList = blockedNicknames.isEmpty ? "blocked peers (not currently online)" : blockedNicknames.sorted().joined(separator: ", ")
-                    let systemMessage = BitchatMessage(
-                        sender: "system",
-                        content: "blocked peers: \(blockedList)",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(systemMessage)
-                }
-            }
-            
-        case "/unblock":
-            if parts.count > 1 {
-                let targetName = String(parts[1])
-                // Remove @ if present
-                let nickname = targetName.hasPrefix("@") ? String(targetName.dropFirst()) : targetName
-                
-                // Find peer ID for this nickname
-                if let peerID = getPeerIDForNickname(nickname) {
-                    // Get fingerprint
-                    if let fingerprintStr = meshService.getPeerFingerprint(peerID) {
-                        
-                        if SecureIdentityStateManager.shared.isBlocked(fingerprint: fingerprintStr) {
-                            // Update social identity to unblock
-                            SecureIdentityStateManager.shared.setBlocked(fingerprintStr, isBlocked: false)
-                            
-                            // Update local set for UI
-                            blockedUsers.remove(fingerprintStr)
-                            
-                            let systemMessage = BitchatMessage(
-                                sender: "system",
-                                content: "unblocked \(nickname).",
-                                timestamp: Date(),
-                                isRelay: false
-                            )
-                            messages.append(systemMessage)
-                        } else {
-                            let systemMessage = BitchatMessage(
-                                sender: "system",
-                                content: "\(nickname) is not blocked.",
-                                timestamp: Date(),
-                                isRelay: false
-                            )
-                            messages.append(systemMessage)
-                        }
-                    } else {
-                        let systemMessage = BitchatMessage(
-                            sender: "system",
-                            content: "cannot unblock \(nickname): unable to verify identity.",
-                            timestamp: Date(),
-                            isRelay: false
-                        )
-                        messages.append(systemMessage)
-                    }
-                } else {
-                    let systemMessage = BitchatMessage(
-                        sender: "system",
-                        content: "cannot unblock \(nickname): user not found.",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(systemMessage)
-                }
-            } else {
-                let systemMessage = BitchatMessage(
-                    sender: "system",
-                    content: "usage: /unblock <nickname>",
-                    timestamp: Date(),
-                    isRelay: false
-                )
-                messages.append(systemMessage)
-            }
-            
-        case "/fav":
-            if parts.count > 1 {
-                let targetName = String(parts[1])
-                // Remove @ if present
-                let nickname = targetName.hasPrefix("@") ? String(targetName.dropFirst()) : targetName
-                
-                // Find peer ID for this nickname
-                if let peerID = getPeerIDForNickname(nickname) {
-                    // Add to favorites using the Nostr integration
-                    if let noisePublicKey = Data(hexString: peerID) {
-                        // Get or set Nostr public key
-                        let existingFavorite = FavoritesPersistenceService.shared.getFavoriteStatus(for: noisePublicKey)
-                        FavoritesPersistenceService.shared.addFavorite(
-                            peerNoisePublicKey: noisePublicKey,
-                            peerNostrPublicKey: existingFavorite?.peerNostrPublicKey,
-                            peerNickname: nickname
-                        )
-                        
-                        // Toggle favorite in identity manager for UI
-                        toggleFavorite(peerID: peerID)
-                        
-                        // Send favorite notification
-                        Task { [weak self] in
-                            try? await self?.messageRouter?.sendFavoriteNotification(to: noisePublicKey, isFavorite: true)
-                        }
-                        
-                        let systemMessage = BitchatMessage(
-                            sender: "system",
-                            content: "added \(nickname) to favorites.",
-                            timestamp: Date(),
-                            isRelay: false
-                        )
-                        messages.append(systemMessage)
-                    }
-                } else {
-                    let systemMessage = BitchatMessage(
-                        sender: "system",
-                        content: "can't find peer: \(nickname)",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(systemMessage)
-                }
-            } else {
-                let systemMessage = BitchatMessage(
-                    sender: "system",
-                    content: "usage: /fav <nickname>",
-                    timestamp: Date(),
-                    isRelay: false
-                )
-                messages.append(systemMessage)
-            }
-            
-        case "/unfav":
-            if parts.count > 1 {
-                let targetName = String(parts[1])
-                // Remove @ if present
-                let nickname = targetName.hasPrefix("@") ? String(targetName.dropFirst()) : targetName
-                
-                // Find peer ID for this nickname
-                if let peerID = getPeerIDForNickname(nickname) {
-                    // Remove from favorites
-                    if let noisePublicKey = Data(hexString: peerID) {
-                        FavoritesPersistenceService.shared.removeFavorite(peerNoisePublicKey: noisePublicKey)
-                        
-                        // Toggle favorite in identity manager for UI
-                        toggleFavorite(peerID: peerID)
-                        
-                        // Send unfavorite notification
-                        Task { [weak self] in
-                            try? await self?.messageRouter?.sendFavoriteNotification(to: noisePublicKey, isFavorite: false)
-                        }
-                        
-                        let systemMessage = BitchatMessage(
-                            sender: "system",
-                            content: "removed \(nickname) from favorites.",
-                            timestamp: Date(),
-                            isRelay: false
-                        )
-                        messages.append(systemMessage)
-                    }
-                } else {
-                    let systemMessage = BitchatMessage(
-                        sender: "system",
-                        content: "can't find peer: \(nickname)",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    messages.append(systemMessage)
-                }
-            } else {
-                let systemMessage = BitchatMessage(
-                    sender: "system",
-                    content: "usage: /unfav <nickname>",
-                    timestamp: Date(),
-                    isRelay: false
-                )
-                messages.append(systemMessage)
-            }
-            
-        case "/testnostr":
-            let systemMessage = BitchatMessage(
-                sender: "system",
-                content: "testing nostr relay connectivity...",
-                timestamp: Date(),
-                isRelay: false
-            )
-            messages.append(systemMessage)
-            
-            Task { @MainActor in
-                if let relayManager = self.nostrRelayManager {
-                    // Simple connectivity test
-                    relayManager.connect()
-                    
-                    // Wait a moment for connections
-                    try? await Task.sleep(nanoseconds: 2_000_000_000) // 2 seconds
-                    
-                    let statusMessage = if relayManager.isConnected {
-                        "nostr relays connected successfully!"
-                    } else {
-                        "failed to connect to nostr relays - check console for details"
-                    }
-                    
-                    let completeMessage = BitchatMessage(
-                        sender: "system",
-                        content: statusMessage,
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    self.messages.append(completeMessage)
-                } else {
-                    let errorMessage = BitchatMessage(
-                        sender: "system",
-                        content: "nostr relay manager not initialized",
-                        timestamp: Date(),
-                        isRelay: false
-                    )
-                    self.messages.append(errorMessage)
-                }
-            }
-            
-        default:
-            // Unknown command
-            let systemMessage = BitchatMessage(
-                sender: "system",
-                content: "unknown command: \(cmd).",
-                timestamp: Date(),
-                isRelay: false
-            )
-            messages.append(systemMessage)
-        }
+        commandProcessor.handle(command)
     }
-    
     // MARK: - Message Reception
     
     func handleHandshakeRequest(from peerID: String, nickname: String, pendingCount: UInt8) {
